consolidate autolabel_shuffle2 and make_shuffle2_reorder() into shuffle_tools
